+++ todo/src/components/card/Card.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import Button from "../button/Button";
 
 const Card = ({ todoText, taskId, status, doneTask, deleteTask }) => (
-  <li className="block py-4 shadow-lg bg-blue-300 px-8 rounded-sm mb-4 flex justify-between items-center gap-12 w-full max-w-4xl group hover:bg-blue-400">
+  <li className="py-4 shadow-lg bg-blue-300 px-8 rounded-sm mb-4 flex justify-between items-center gap-12 w-full max-w-4xl group hover:bg-blue-400">
     <div className={`flex ${status ? "line-through" : ""}`}>
       <p className="text-xl">{taskId + 1}.</p>
       <div className="pl-6 text-xl">{todoText}</div>
